Update BasketballStar.wp
The was an issue regarding the basketball star game, where the player does not earn a score when the basketball is in the hoop. Furthermore, the player sometimes earns a score even if the basketball is not in the hoop. Through trial and error, I changed the range of the basketXPlace and included the y position of the basketball in consideration of the score, since the ball can sometimes drift away from the basket as it goes down. An issue I ran into was that sometimes, the score wouldn't go up if the ball touches the rim and rolls into the basket.