Create ProjectPreview.svelte
Issue 606 Question:
I am a little concerned with my work and would like some help. I am using uid to find id with Creator, while it seems that it cannot find the uid. Did I find id in the wrong place or is it because of limited permission? I have also written the codes for collaborated authors, while I commented it out for now because I am still having issue with single author.
This is the path that I downloaded the codes: wordplay\src\components\project. Did I get the issue because of await Creators.getCreator(owner), which might require a longer time to get the username? I will continue to figure it out and test myself. I understand that you are busy, while could you please also see if my codes work on your end if possible? This is the path to my file.